+++ api/controllers/console/app/agent.py
@@ -7,10 +7,10 @@ from controllers.console import console_ns
 from controllers.console.app.wraps import get_app_model
 from controllers.console.wraps import (
     account_initialization_required,
-    rbac_permission_required,
     setup_required,
     with_current_user,
 )
+from controllers.common.wraps import rbac_permission_required
 from extensions.ext_database import db
 from libs.helper import uuid_value
 from libs.login import login_required
