+++ api/controllers/console/app/agent_app_access.py
@@ -13,10 +13,10 @@ from controllers.console import console_ns
 from controllers.console.app.wraps import get_app_model
 from controllers.console.wraps import (
     account_initialization_required,
-    rbac_permission_required,
     setup_required,
     with_current_tenant_id,
 )
+from controllers.common.wraps import rbac_permission_required
 from extensions.ext_database import db
 from fields.base import ResponseModel
 from libs.login import login_required
