+++ api/controllers/console/app/workflow_statistic.py
@@ -8,10 +8,10 @@ from controllers.console import console_ns
 from controllers.console.app.wraps import get_app_model
 from controllers.console.wraps import (
     account_initialization_required,
-    rbac_permission_required,
     setup_required,
     with_current_user,
 )
+from controllers.common.wraps import rbac_permission_required
 from extensions.ext_database import db
 from libs.datetime_utils import parse_time_range
 from libs.login import login_required
