+++ api/controllers/console/app/workflow_trigger.py
@@ -22,10 +22,10 @@ from ..app.wraps import get_app_model
 from ..wraps import (
     account_initialization_required,
     edit_permission_required,
-    rbac_permission_required,
     setup_required,
     with_current_tenant_id,
 )
+from controllers.common.wraps import rbac_permission_required
 
 logger = logging.getLogger(__name__)
