+++ api/tests/unit_tests/controllers/openapi/auth/test_step_authz.py
@@ -2,11 +2,11 @@ from types import SimpleNamespace
 from unittest.mock import patch
 
 import pytest
-from werkzeug.exceptions import Forbidden
-
 from controllers.openapi.auth.context import Context
 from controllers.openapi.auth.steps import AppAuthzCheck
 from controllers.openapi.auth.strategies import AclStrategy, MembershipStrategy
+from werkzeug.exceptions import Forbidden
+
 from libs.oauth_bearer import SubjectType
